Renaming tenants in misc files and fixing up filenames
