+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Models/OrganizationSummaryModel.cs
@@ -1,0 +1,10 @@
+﻿namespace AllReady.Areas.Admin.Models
+{
+    public class OrganizationSummaryModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string LogoUrl { get; set; }
+        public string WebUrl { get; set; }
+    }
+}
