--- AllReadyApp/Web-App/AllReady/Areas/Admin/Models/TenantSummaryModel.cs
@@ -1,10 +1,0 @@
-﻿namespace AllReady.Areas.Admin.Models
-{
-    public class OrganizationSummaryModel
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string LogoUrl { get; set; }
-        public string WebUrl { get; set; }
-    }
-}
