+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Organization/Delete.cshtml
@@ -1,4 +1,4 @@
-@model AllReady.Models.Organization
+@model AllReady.Areas.Admin.Models.OrganizationDetailModel
 
 @{
     ViewData["Title"] = "Delete " + Model.Name;
