+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Organization/Index.cshtml
@@ -1,0 +1,54 @@
+@model IEnumerable<AllReady.Areas.Admin.Models.OrganizationSummaryModel>
+
+@{
+    ViewData["Title"] = "Currently active organizations";
+}
+
+<doiv class="row">
+    <div class="col-md-12">
+        <h2>@ViewData["Title"]</h2>
+    </div>
+</doiv>
+<div class="row">
+    <div class="col-md-12">
+        <table class="table">
+            <tr>
+                <th>
+                    @Html.DisplayNameFor(model => model.Name)
+                </th>
+                <th>
+                    @Html.DisplayNameFor(model => model.WebUrl)
+                </th>
+                <th></th>
+            </tr>
+            @foreach (var item in Model) {
+                <tr>
+                    <td>
+                        <a href="@Url.Action("Details", "Organization", new { id = item.Id })">
+                        @if (String.IsNullOrEmpty(item.LogoUrl))
+                        {
+                            @item.Name
+                        }
+                        else
+                        {
+                            <img src="@item.LogoUrl" alt="@item.Name" />
+                        }
+                        </a>
+                    </td>
+                    <td>
+                        @Html.DisplayFor(modelItem => item.WebUrl)
+                    </td>
+                    <td>
+                        <a href="@Url.Action("Edit", "Organization", new { id = item.Id })" class="btn btn-default btn-sm" ><span class="fa fa-edit"></span> Edit</a>
+                        <a href="@Url.Action("Delete", "Organization", new { id = item.Id })" class="btn btn-default btn-sm"><span class="fa fa-trash"></span> Delete</a>
+                    </td>
+                </tr>
+            }
+        </table>
+    </div>
+</div>
+<div class="row">
+    <div class="col-md-12">
+        @Html.ActionLink("Create New", "Create", "Organization", null, new { @class = "btn" })
+    </div>
+</div>
