+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Site/RevokeOrganizationAdmin.cshtml
@@ -1,7 +1,7 @@
 ﻿@{
     ViewBag.Title = "Revoke Organization Admin";
 }
-<h2>Revoke Site Admin</h2>
+<h2>Revoke Organization Admin</h2>
 @if (ViewBag.ErrorMessage != null)
 {
     <div class="alert alert-danger">
