--- AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Tenant/Create.cshtml
@@ -1,98 +1,0 @@
-@model AllReady.Areas.Admin.Models.OrganizationEditModel
-
-@{
-    ViewData["Title"] = "Create a new organization";
-}
-<div class="row">
-    <div class="col-md-12">
-        <h2>@ViewData["Title"]</h2>
-    </div>
-</div>
-<form asp-controller="Organization" asp-route-area="Admin" asp-action="Create" method="post">
-    <div class="row">
-        <div class="form-horizontal col-lg-6 pull-left">
-            <h4>Organization</h4>
-            <hr />
-            <div asp-validation-summary="ValidationSummary.ModelOnly" class="text-danger"></div>
-
-            <div class="form-group">
-
-                <label asp-for="LogoUrl" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="LogoUrl" class="form-control" />
-                    <span asp-validation-for="LogoUrl" class="text-danger"></span>
-                </div>
-
-            </div>
-            <div class="form-group">
-
-                <label asp-for="Name" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Name" class="form-control" />
-                    <span asp-validation-for="Name" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group">
-
-                <label asp-for="WebUrl" class="col-md-2 control-label">Web Url</label>
-                <div class="col-md-10">
-                    <input asp-for="WebUrl" class="form-control" />
-                    <span asp-validation-for="WebUrl" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group">
-
-                <label asp-for="Location" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="Location" class="form-control" />
-                    <span asp-validation-for="Location" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="PrimaryContactFirstName" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="PrimaryContactFirstName" class="form-control" />
-                    <span asp-validation-for="PrimaryContactFirstName" class="text-danger"></span>
-                </div>
-
-            </div>
-
-
-            <div class="form-group">
-
-                <label asp-for="PrimaryContactLastName" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="PrimaryContactLastName" class="form-control" />
-                    <span asp-validation-for="PrimaryContactLastName" class="text-danger"></span>
-                </div>
-            </div>
-
-
-
-            <div class="form-group">
-                <label asp-for="PrimaryContactPhoneNumber" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="PrimaryContactPhoneNumber" class="form-control" />
-                    <span asp-validation-for="PrimaryContactPhoneNumber" class="text-danger"></span>
-                </div>
-            </div>
-
-            <div class="form-group">
-                <label asp-for="PrimaryContactEmail" class="col-md-2 control-label"></label>
-                <div class="col-md-10">
-                    <input asp-for="PrimaryContactEmail" class="form-control" />
-                    <span asp-validation-for="PrimaryContactEmail" class="text-danger"></span>
-                </div>
-            </div>
-            <div class="form-group">
-                <div class="col-md-offset-2 col-md-10">
-                    <button type="submit" class="btn btn-default">Save</button>
-                    <a asp-controller="Organization" asp-action="Index" class="btn btn-cancel" />
-                </div>
-            </div>
-        </div>
-    </div>
-</form>
-
