+++ AllReadyApp/Web-App/AllReady/Views/Organization/Index.cshtml
@@ -11,7 +11,7 @@
         <div class="pull-right">
             <div class="form-inline">
                 <div class="form-group">
-                    <input type="text" class="form-control input-sm" data-bind="textInput: organization.searchTerm" placeholder="Search organizations" />
+                    <input type="text" class="form-control input-sm" data-bind="textInput: organizations.searchTerm" placeholder="Search organizations" />
                 </div>
                 <span data-bind="text: '(' + total + ')' "></span>
             </div>
